PP-6325: export page button on payment links
