+++ src/web/modules/payment_links/overview.njk
@@ -38,6 +38,18 @@
   {% endif %}
 
   <div>
+    {% set csvHref =
+      ("/gateway_accounts/" + accountId + "/payment_links/csv?live=" + filterLiveAccounts + "&sort=" + sort)
+      if accountId else
+      ("/payment_links/csv?live=" + filterLiveAccounts + "&sort=" + sort) %}
+    <a href="{{ csvHref }}"
+      class="govuk-button govuk-button--secondary govuk-!-margin-bottom-2">
+      Export page as CSV
+    </a>
+  </div>
+  <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
+
+  <div>
     <table class="govuk-table">
       <thead class="govuk-table__head">
         <tr class="govuk-table__row">
@@ -83,8 +95,7 @@
           </tr>
           <tr class="govuk-table__row">
             <td class="govuk-table__cell" colspan="4">
-              {# {% set linkHref = link.product._indexedLinks.friendly or link.product._indexedLinks.pay %} #}
-              <a class="govuk-link govuk-link--no-visited-state" href="{{ linkHref }}">
+              <a class="govuk-link govuk-link--no-visited-state" href="{{ link.url }}">
                 {{ link.url | truncate(80) }}
               </a>
           </tr>
