+++ src/web/modules/payment_links/payment_links.http.ts
@@ -1,8 +1,7 @@
 import _ from 'lodash'
 import { Request, Response, NextFunction } from 'express'
 import { Products, AdminUsers, Connector } from '../../../lib/pay-request'
-import { Parser } from 'json2csv'
-
+import { format } from './csv'
 function indexPaymentLinksByType(paymentLink: any): any {
   const index = paymentLink.product._links.reduce((aggregate: any, linkDetails: any) => {
     aggregate[linkDetails.rel] = linkDetails.href
@@ -40,79 +39,18 @@ export async function listCSV(req: Request, res: Response, next: NextFunction): 
     const { accountId } = req.params
     const sort = req.query.sort || 'last_payment_date'
     const live = req.query.live === undefined ? true : req.query.live && req.query.live !== "false"
-
-    // we're only going to filter live gateway accounts if we're at the whole platform level
     const filterLiveAccounts = live && !accountId
     const usageReportResults = await fetchUsageContext(sort, filterLiveAccounts, accountId)
-
     const flatLinksList = usageReportResults.groupedPaymentLinks
       .reduce((aggregate: any, groupedList: any) => {
         aggregate = [ ...aggregate, ...groupedList.links ]
         return aggregate
       }, [])
 
-    const fields = [{
-      label: 'ID',
-      value: 'product.external_id'
-    }, {
-      label: 'Gateway account is live',
-      value: 'is_live_account'
-    }, {
-      label: 'Gateway account ID',
-      value: 'product.gateway_account_id'
-    }, {
-      label: 'Number of payments',
-      value: 'payment_count'
-    }, {
-      label: 'Date last used',
-      value: 'last_payment_date'
-    },{
-      label: 'Payment link name',
-      value: 'product.name'
-    }, {
-      label: 'Payment link description',
-      value: 'product.description'
-    },{
-      label: 'Payment link URL',
-      value: 'url'
-    },{
-      label: 'Service name',
-      value: 'service_name'
-    },{
-      label: 'Organisation name',
-      value: 'organisation_name'
-    },{
-      label: 'Sector',
-      value: 'sector'
-    },{
-      label: 'Is fixed price',
-      value: 'is_fixed_price'
-    },{
-      label: 'Fixed price',
-      value: 'product.price'
-    },{
-      label: 'Custom reference entered',
-      value: 'product.reference_enabled'
-    },{
-      label: 'Reference label',
-      value: 'product.reference_label'
-    },{
-      label: 'Language',
-      value: 'product.language'
-    },{
-      label: 'Has metadata',
-      value: 'has_metadata'
-    },{
-      label: 'Metadata',
-      value: 'product.metadata'
-    }]
-
-    const parser = new Parser({ fields })
-    const csv = parser.parse(flatLinksList)
     const name = accountId || 'platform'
     res.set('Content-Type', 'text/csv')
     res.set('Content-Disposition', `attachment; filename="GOVUK_Pay_payment_links_usage_${ name }.csv"`)
-    res.status(200).send(csv)
+    res.status(200).send(format(flatLinksList))
    } catch(error) {
      next(error)
    }
